Move leetcode tests to it's own folder
