+++ src/test/java/com/patina/codebloom/leetcode/LeetcodeApiHandlerTest.java
@@ -1,4 +1,4 @@
-package com.patina.codebloom;
+package com.patina.codebloom.leetcode;
 
 import java.util.ArrayList;
