+++ src/test/java/com/patina/codebloom/leetcode/ScoreCalculatorTest.java
@@ -1,4 +1,4 @@
-package com.patina.codebloom;
+package com.patina.codebloom.leetcode;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
